Reprise a 0

+++ libft/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: benjamin <benjamin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/07 16:57:13 by btollet           #+#    #+#             */
-/*   Updated: 2016/12/18 17:12:20 by benjamin         ###   ########.fr       */
+/*   Updated: 2016/12/19 14:32:25 by benjamin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ char	*ft_strdup(const char *s)
 {
 	char	*result;
 
-	result = (char*)malloc(sizeof(char) * (int)ft_strlen(s) + 1);
+	result = (char*)malloc(sizeof(char) * ft_strlen(s) + 1);
 	if (!result)
 		return (NULL);
 	ft_strcpy(result, s);
